requirements.txt
Django==1.5
Fabric==1.5.3
South==0.7.6
django-debug-toolbar==0.9.4
raven==3.1.16
dj-database-url==0.2.1
gunicorn==0.17.2
django-nose==1.1
coverage==3.6
nose==1.2.1
nosexcover==1.0.7
django-compressor==1.2
pyOpenSSL==0.13
